data_lake/ngs-update-metadata/container/requirements.txt
azure-core==1.18.0
azure-functions==1.7.2
azure-identity==1.6.1
azure-storage-blob==12.8.1
psycopg2-binary==2.9.1
argparse==1.4.0
coloredlogs==15.0.1
black==21.8b0